Add a new component to reduce the number of wands
Add a new component called "Reduce the number of wands". This component
limits the number of wands that drop and are sold in shops. Additionaly,
it also disables the somewhat questionable recharging mechanism, i.e.
being able to sell and re-buy the same wand with the maximum amount of
charges.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>